Cluster checks table makeover: achieve a hierarchical view
