+++ web/frontend/javascripts/check_results.js
@@ -2,23 +2,9 @@ import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { get } from 'axios';
 
-import Accordion from '@components/Accordion';
 import ChecksTable from '@components/ChecksTable';
 
 const clusterId = window.location.pathname.split('/').pop();
-
-const groupChecks = (checks) => {
-  const groups = checks.reduce((accumulator, current) => {
-    const { group } = current;
-    return accumulator[group]
-      ? { ...accumulator, [group]: [...accumulator[group], current] }
-      : { ...accumulator, [group]: [current] };
-  }, {});
-
-  return Object.keys(groups).map((key) => {
-    return { name: key, checks: groups[key] };
-  });
-};
 
 const ClustersChecks = ({ clusterId }) => {
   const [results, setResults] = useState([]);
@@ -26,25 +12,14 @@ const ClustersChecks = ({ clusterId }) => {
 
   useEffect(() => {
     get(`/api/clusters/${clusterId}/results`).then(({ data }) => {
-      const groupedChecks = groupChecks(data.checks);
-      setResults(groupedChecks);
+      setResults(data.checks);
       setHosts(data.hosts);
     });
   }, []);
 
   return (
     <div>
-      {results.map((section) => {
-        return (
-          <Accordion
-            className="checks-results-accordion"
-            key={section.name}
-            title={section.name}
-          >
-            <ChecksTable checks={section.checks} clusterHosts={hosts} />
-          </Accordion>
-        );
-      })}
+      <ChecksTable checks={results} clusterHosts={hosts} />
     </div>
   );
 };
